@@ -6,11 +6,13 @@
     "homepage": "https://github.com/twilio/authy-php",
     "license": "MIT",
     "require": {
-        "php": ">=5.6.0",
-        "guzzlehttp/guzzle": "~6.0"
+        "php": ">=7.4",
+        "guzzlehttp/guzzle": "~6.0",
+        "ext-curl": "*",
+        "ext-json": "*"
     },
     "require-dev": {
-        "EHER/PHPUnit": ">= 1.6"
+        "phpunit/phpunit": "^9.4.0"
     },
     "authors": [
         {
